Trigger rebuilds on schedule
Redeploy the site every morning to ensure it has an up-to-date copy of schemas from opg-data-lpa-store.

Also make it manually dispatchable so that we can trigger it if we need an urgent update.

This is an alternative to remote repository dispatches, which would have required personal access tokens.

Fixes CTC-149 #patch